pyproject.toml
[project]
name = "somos-juntos-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.6.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "bandit>=1.7.10",
    "black>=22.1.0",
    "blue>=0.9.1",
    "flake8>=4.0.1",
    "isort>=5.13.2",
    "mypy>=1.12.1",
    "pip-audit>=2.7.3",
    "pre-commit>=4.0.1",
    "pydocstyle>=6.3.0",
    "pylint>=3.3.1",
    "taskipy>=1.13.0",
]
